Create FaceRecog.py
This is a Face recognition attendance system where the code recognizes the face captured in frame if the matching face images is present in folder. If so, then the person's name is displayed in frame and also his name and time when face was recognized is registered in csv file.